LaTeX writer: output H1 as "part" unless "book" or "chapters" option is provided
This change eliminates errors when rendering a plain document with a level 1 header to LaTeX, due to the previous mapping to "chapter" not available in non-book documentclasses.

This adds a new CLI option "book" that, when set, automatically maps H1 to "chapters", and enables chapters/parts numbering.

Requires new Scholmd LaTeX template to suppress part numbering in article-style documentclasses.