Added "--force-crossrefs" option to CLI
Passing this will force all parsing of cross-references to ignore the safety check of whether a corresponding identifier exists in the document.

This may be useful for individually compiling LaTeX for inclusion in other documents, or for cases where numeric labels would later be filled-in with a filter.

If a corresponding numeric label is not found, the default label text is replaced by the raw text. This will not affect LaTeX.